fix: replace a components render with a clone of its fragment
Inject a clone of the rendered fragment instead of the original
fragment. This way the original fragment can be potentially cached and
returned in multiple renders from multiple components or the same
component over multiple renders.

This also helps with replace the "same" DOM. While replace the same DOM
with identical nodes, it isn't the original nodes themselves.